+++ python/silodb/__init__.py
@@ -1,4 +1,4 @@
-"""PySilo - Python bindings for Silo"""
+"""silodb - Python bindings for SILO"""
 
 from importlib.metadata import version, PackageNotFoundError
 
@@ -7,6 +7,6 @@ from .database import PyDatabase as Database
 __all__ = ['Database']
 
 try:
-    __version__ = version("pysilo")
+    __version__ = version("silodb")
 except PackageNotFoundError:
     __version__ = "0.0.0"  # Fallback for development