+++ python/silodb/database.pyx
@@ -1,4 +1,4 @@
-# python/pysilo/database.pyx
+# python/silodb/database.pyx
 from libcpp.string cimport string
 from libcpp.vector cimport vector
 from libcpp.pair cimport pair